requirements.txt
flask==2.3.3
pandas==2.1.0
numpy==1.25.2
matplotlib==3.7.3
openai==1.65.3
python-pptx==0.6.21
reportlab==4.0.4
python-dotenv==1.0.0
gunicorn==21.2.0
seaborn==0.13.2
xhtml2pdf==0.2.17
markdown==3.4.3